Trait assosiation type.
