+++ examples/l10_trait/src/assosiation/entities.rs
@@ -1,0 +1,8 @@
+use serde::{Serialize, Deserialize};
+
+#[derive(Serialize, Deserialize, Debug, Default)]
+pub struct Product{
+    pub id: String,
+    pub name: String,
+    pub price: u32,
+}
