+++ examples/l10_trait/src/assosiation/mod.rs
