+++ examples/l10_trait/src/assosiation/traits_impl.rs
@@ -8,17 +8,19 @@ use std::{marker::PhantomData, path::PathBuf, fs::read_to_string, fs::File, io::
 pub struct CsvReaderImpl<T> {
     pub phantom: PhantomData<T>,
 }
-impl<T> CsvReader<T> for CsvReaderImpl<T>
+impl<T> CsvReader for CsvReaderImpl<T>
 where
     T: DeserializeOwned,
 {
-    fn read(&self, file_path: &str) -> Result<Vec<T>> {
+    type Entity = T;
+
+    fn read(&self, file_path: &str) -> Result<Vec<Self::Entity>> {
         let path_buff = PathBuf::from(file_path);
 
         let string_data = read_to_string(path_buff)?;
         let mut reader = ReaderBuilder::new().delimiter(b',').from_reader(string_data.as_bytes());
-        let rows = reader.deserialize::<T>();
-        let mut result = Vec::<T>::new();
+        let rows = reader.deserialize::<Self::Entity>();
+        let mut result = Vec::<Self::Entity>::new();
         for row in rows {
             result.push(row?);
         }
@@ -29,14 +31,16 @@ where
 pub struct JsonReaderImpl<T> {
     pub phantom: PhantomData<T>,
 }
-impl<T> JsonReader<T> for JsonReaderImpl<T>
+impl<T> JsonReader for JsonReaderImpl<T>
 where
     T: DeserializeOwned,
 {
-    fn read(&self, file_path: &str) -> Result<Vec<T>> {
+    type Entity = T;
+
+    fn read(&self, file_path: &str) -> Result<Vec<Self::Entity>> {
         let path_buff = PathBuf::from(file_path);
         let buf_reader = File::open(path_buff).map(BufReader::new)?;
-        let result = serde_json::from_reader::<BufReader<File>, Vec<T>>(buf_reader)?;
+        let result = serde_json::from_reader::<BufReader<File>, Vec<Self::Entity>>(buf_reader)?;
         Ok(result)
     }
 }
