+++ examples/l10_trait/src/lib.rs
@@ -1,2 +1,3 @@
 mod basic;
 pub(crate) mod generics;
+pub(crate) mod assosiation;
